pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pg_freezer"]

[project]
name = "pg-freezer"
version = "0.1.0"
description = "Archive PostgreSQL timeseries tables to Parquet on S3-compatible storage"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.29",
    "pyarrow>=15.0",
    "fsspec>=2024.2",
    "s3fs>=2024.2",
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "click>=8.1",
    "structlog>=24.1",
    "httpx>=0.27",
]

[project.scripts]
pg-freezer = "pg_freezer.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "duckdb>=0.10",
    "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "duckdb>=0.10",
    "pytest-cov>=5.0",
]
